test(integration): move KafkaSQL startup verification into the IT matrix
utils/extra-tests existed as a separate CI job running a single IT class
(KafkaSqlStartupVerificationIT) and was the extras critical path at ~21 min.
Move the test and its infra into integration-tests as the kafkasql-startup
group so it runs as a regular matrix entry (storage=kafkasql,
remote-kafka), parallel with everything else:

- test + infra moved to io.apicurio.tests.kafkasql.startup(.infra),
  tagged with the new kafkasql-startup constant
- matrix entry added; -Dregistry3-image wired so the freshly built image
  is tested (v2 image still defaults to the quay latest-release)
- AssertJ assertions converted to JUnit (integration-tests house style)
- utils/extra-tests module removed from the reactor and .fossa.yml
- the old extra-tests maven step removed from verify-extras.yaml

Verified locally: test compiles in its new home and a negative case runs
green via -Dgroups=kafkasql-startup in ~10 s (previously ~120 s+ before
the fast-failure fix in the parent commit).
